make * not jump to the next occurence, but still hlsearch the current word
reasoning: we already have a plugin that highlights other occurrences of
the current word as soon as the cursor is placed on a word, but it's
bound to the current lexical scope. sometimes we want to preview and
highlight all occurrences of the word no matter where they are on screen
or in the file, but keep the cursor where it is.